refactor: 게임참여 Dto를 Entity로 변경, 내부 필드 타입 수정
