+++ src/main/java/io/cavia/trader/module/member/repository/GameParticipationMapper.java
@@ -1,6 +1,6 @@
 package io.cavia.trader.module.member.repository;
 
-import io.cavia.trader.module.member.dto.GameParticipationDto;
+import io.cavia.trader.module.member.entity.GameParticipation;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
@@ -8,5 +8,5 @@ import java.util.List;
 @Mapper
 public interface GameParticipationMapper {
 
-    List<GameParticipationDto> findByMemberId(Long memberId);
+    List<GameParticipation> findByMemberId(Long memberId);
 }
