+++ src/main/java/io/cavia/trader/module/member/service/MemberService.java
@@ -1,6 +1,6 @@
 package io.cavia.trader.module.member.service;
 
-import io.cavia.trader.module.member.dto.GameParticipationDto;
+import io.cavia.trader.module.member.entity.GameParticipation;
 import io.cavia.trader.module.member.dto.UserRankingDto;
 import io.cavia.trader.module.member.dto.PasswordChangeRequestDto;
 import io.cavia.trader.module.member.entity.Member;
@@ -9,7 +9,7 @@ import java.util.List;
 
 public interface MemberService {
 
-    List<GameParticipationDto> getGameParticipationByMemberId(Long memberId);
+    List<GameParticipation> getGameParticipationByMemberId(Long memberId);
 
     Member getMemberById(Long id);
